Make db, model, prefs, recycler, repo, ui, vm packages. Move files to respective packages.
