+++ M01/Assignment/NoteTakerRoom/app/src/main/java/com/lambdaschool/notetakerroom/db/NotesRoomDao.kt
@@ -1,4 +1,4 @@
-package com.lambdaschool.notetakerroom
+package com.lambdaschool.notetakerroom.db
 
 import androidx.room.Dao
 import androidx.room.Insert
