+++ M01/Assignment/NoteTakerRoom/app/src/main/java/com/lambdaschool/notetakerroom/model/Note.kt
